Move compile_commands fixture to conftest.py file
